281: move tests in interpret/ to proper locations
