+++ src/global/test/utils.spec.js
@@ -1,6 +1,7 @@
 import test from 'tape';
 
-import { assert, assertGridInBounds } from '../utils';
+import { assert, assertGridInBounds, pointsOnSameSide } from '../utils';
+import { Point } from '../../interpret/class/Point';
 
 
 test('assert: does not throw errors with condition=true', t => {
@@ -35,4 +36,38 @@ test('assertGridInBounds: throws errors when inputs are not in bounds', t => {
   t.throws(() => { assertGridInBounds(arr, 1, 4); });
 
   t.end();
+});
+
+
+test('pointsOnSameSide()', tester => {
+  tester.test('returns true when points are on the left side of the edge', t => {
+    const edge = { p1: new Point(3, 0), p2: new Point(3, 1) };
+    const p1 = new Point(0, 0);
+    const p2 = new Point(0, 1);
+
+    t.true(pointsOnSameSide(p1, p2, edge));
+
+    t.end();
+  });
+
+  tester.test('returns true when points are on the right side of the edge', t => {
+    const edge = { p1: new Point(3, 0), p2: new Point(3, 1) };
+    const p1 = new Point(4, 0);
+    const p2 = new Point(4, 1);
+
+    t.true(pointsOnSameSide(p1, p2, edge));
+
+    t.end();
+  });
+
+  tester.test('returns false when points are on opposite sides of the edge', t => {
+    const edge = { p1: new Point(3, 0), p2: new Point(3, 1) };
+    const p1 = new Point(0, 0);
+    const p2 = new Point(4, 1);
+
+    t.false(pointsOnSameSide(p1, p2, edge));
+
+    t.end();
+  });
+  tester.end();
 });
